[FLINK-1945][py] Python Tests less verbose
Tests now extend from JavaProgramTestBase, and were consolidated into less files to reduce build time.
