+++ flink-libraries/flink-python/src/test/python/org/apache/flink/python/api/test_main.py
@@ -21,107 +21,11 @@ from flink.functions.FlatMapFunction import FlatMapFunction
 from flink.functions.FilterFunction import FilterFunction
 from flink.functions.MapPartitionFunction import MapPartitionFunction
 from flink.functions.ReduceFunction import ReduceFunction
-from flink.functions.CrossFunction import CrossFunction
-from flink.functions.JoinFunction import JoinFunction
 from flink.functions.GroupReduceFunction import GroupReduceFunction
-from flink.functions.CoGroupFunction import CoGroupFunction
-from flink.plan.Constants import INT, STRING, FLOAT, BOOL, CUSTOM, Order
+from flink.plan.Constants import INT, STRING, FLOAT, BOOL, BYTES, CUSTOM, Order, WriteMode
 import struct
 
-
-class Mapper(MapFunction):
-    def map(self, value):
-        return value * value
-
-
-class Filter(FilterFunction):
-    def __init__(self, limit):
-        super(Filter, self).__init__()
-        self.limit = limit
-
-    def filter(self, value):
-        return value > self.limit
-
-
-class FlatMap(FlatMapFunction):
-    def flat_map(self, value, collector):
-        collector.collect(value)
-        collector.collect(value * 2)
-
-
-class MapPartition(MapPartitionFunction):
-    def map_partition(self, iterator, collector):
-        for value in iterator:
-            collector.collect(value * 2)
-
-
-class Reduce(ReduceFunction):
-    def reduce(self, value1, value2):
-        return value1 + value2
-
-
-class Reduce2(ReduceFunction):
-    def reduce(self, value1, value2):
-        return (value1[0] + value2[0], value1[1] + value2[1], value1[2], value1[3] or value2[3])
-
-
-class Cross(CrossFunction):
-    def cross(self, value1, value2):
-        return (value1, value2[3])
-
-
-class MapperBcv(MapFunction):
-    def map(self, value):
-        factor = self.context.get_broadcast_variable("test")[0][0]
-        return value * factor
-
-
-class Join(JoinFunction):
-    def join(self, value1, value2):
-        if value1[3]:
-            return value2[0] + str(value1[0])
-        else:
-            return value2[0] + str(value1[1])
-
-
-class GroupReduce(GroupReduceFunction):
-    def reduce(self, iterator, collector):
-        if iterator.has_next():
-            i, f, s, b = iterator.next()
-            for value in iterator:
-                i += value[0]
-                f += value[1]
-                b |= value[3]
-            collector.collect((i, f, s, b))
-
-
-class GroupReduce2(GroupReduceFunction):
-    def reduce(self, iterator, collector):
-        for value in iterator:
-            collector.collect(value)
-
-
-class GroupReduce3(GroupReduceFunction):
-    def reduce(self, iterator, collector):
-        collector.collect(iterator.next())
-
-    def combine(self, iterator, collector):
-        if iterator.has_next():
-            v1 = iterator.next()
-        if iterator.has_next():
-            v2 = iterator.next()
-        if v1[0] < v2[0]:
-            collector.collect(v1)
-        else:
-            collector.collect(v2)
-
-
-class CoGroup(CoGroupFunction):
-    def co_group(self, iterator1, iterator2, collector):
-        while iterator1.has_next() and iterator2.has_next():
-            collector.collect((iterator1.next(), iterator2.next()))
-
-
+#Utilities
 class Id(MapFunction):
     def map(self, value):
         return value
@@ -137,10 +41,9 @@ class Verify(MapPartitionFunction):
         index = 0
         for value in iterator:
             if value != self.expected[index]:
-                print(self.name + " Test failed. Expected: " + str(self.expected[index]) + " Actual: " + str(value))
-                raise Exception(self.name + " failed!")
+                raise Exception(self.name + " Test failed. Expected: " + str(self.expected[index]) + " Actual: " + str(value))
             index += 1
-        collector.collect(self.name + " successful!")
+        #collector.collect(self.name + " successful!")
 
 
 class Verify2(MapPartitionFunction):
@@ -155,8 +58,8 @@ class Verify2(MapPartitionFunction):
                 try:
                     self.expected.remove(value)
                 except Exception:
-                    raise Exception(self.name + " failed!")
-        collector.collect(self.name + " successful!")
+                    raise Exception(self.name + " failed! Actual value " + str(value) + "not contained in expected values: "+str(self.expected))
+        #collector.collect(self.name + " successful!")
 
 
 if __name__ == "__main__":
@@ -172,93 +75,33 @@ if __name__ == "__main__":
 
     d5 = env.from_elements((4.4, 4.3, 1), (4.3, 4.4, 1), (4.2, 4.1, 3), (4.1, 4.1, 3))
 
-    d1 \
-        .map((lambda x: x * x), INT).map(Mapper(), INT) \
-        .map_partition(Verify([1, 1296, 20736], "Map"), STRING).output()
+    d6 = env.from_elements(1, 1, 12)
 
-    d1 \
-        .map(Mapper(), INT).map((lambda x: x * x), INT) \
-        .map_partition(Verify([1, 1296, 20736], "Chained Lambda"), STRING).output()
+    #CSV Source/Sink
+    csv_data = env.read_csv("src/test/python/org/apache/flink/python/api/data_csv", (INT, INT, STRING))
 
-    d1 \
-        .filter(Filter(5)).filter(Filter(8)) \
-        .map_partition(Verify([12], "Filter"), STRING).output()
+    csv_data.write_csv("/tmp/flink/result1", line_delimiter="\n", field_delimiter="|", write_mode=WriteMode.OVERWRITE)
 
-    d1 \
-        .flat_map(FlatMap(), INT).flat_map(FlatMap(), INT) \
-        .map_partition(Verify([1, 2, 2, 4, 6, 12, 12, 24, 12, 24, 24, 48], "FlatMap"), STRING).output()
+    #Text Source/Sink
+    text_data = env.read_text("src/test/python/org/apache/flink/python/api/data_text")
 
-    d1 \
-        .map_partition(MapPartition(), INT) \
-        .map_partition(Verify([2, 12, 24], "MapPartition"), STRING).output()
+    text_data.write_text("/tmp/flink/result2", WriteMode.OVERWRITE)
 
-    d1 \
-        .reduce(Reduce()) \
-        .map_partition(Verify([19], "AllReduce"), STRING).output()
+    #Types
+    env.from_elements(bytearray(b"hello"), bytearray(b"world"))\
+        .map(Id(), BYTES).map_partition(Verify([bytearray(b"hello"), bytearray(b"world")], "Byte"), STRING).output()
 
-    d4 \
-        .group_by(2).reduce(Reduce2()) \
-        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "CombineReduce"), STRING).output()
+    env.from_elements(1, 2, 3, 4, 5)\
+        .map(Id(), INT).map_partition(Verify([1,2,3,4,5], "Int"), STRING).output()
 
-    d4 \
-        .map(Id(), (INT, FLOAT, STRING, BOOL)).group_by(2).reduce(Reduce2()) \
-        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "ChainedReduce"), STRING).output()
+    env.from_elements(True, True, False)\
+        .map(Id(), BOOL).map_partition(Verify([True, True, False], "Bool"), STRING).output()
 
-    d1 \
-        .map(MapperBcv(), INT).with_broadcast_set("test", d2) \
-        .map_partition(Verify([1, 6, 12], "Broadcast"), STRING).output()
+    env.from_elements(1.4, 1.7, 12312.23)\
+        .map(Id(), FLOAT).map_partition(Verify([1.4, 1.7, 12312.23], "Float"), STRING).output()
 
-    d1 \
-        .cross(d2).using(Cross(), (INT, BOOL)) \
-        .map_partition(Verify([(1, True), (1, False), (6, True), (6, False), (12, True), (12, False)], "Cross"), STRING).output()
-
-    d1 \
-        .cross(d3) \
-        .map_partition(Verify([(1, ("hello",)), (1, ("world",)), (6, ("hello",)), (6, ("world",)), (12, ("hello",)), (12, ("world",))], "Default Cross"), STRING).output()
-
-    d2 \
-        .cross(d3).project_second(0).project_first(0, 1) \
-        .map_partition(Verify([("hello", 1, 0.5), ("world", 1, 0.5), ("hello", 2, 0.4), ("world", 2, 0.4)], "Project Cross"), STRING).output()
-
-    d2 \
-        .join(d3).where(2).equal_to(0).using(Join(), STRING) \
-        .map_partition(Verify(["hello1", "world0.4"], "Join"), STRING).output()
-
-    d2 \
-        .join(d3).where(2).equal_to(0).project_first(0, 3).project_second(0) \
-        .map_partition(Verify([(1, True, "hello"), (2, False, "world")], "Project Join"), STRING).output()
-
-    d2 \
-        .join(d3).where(2).equal_to(0) \
-        .map_partition(Verify([((1, 0.5, "hello", True), ("hello",)), ((2, 0.4, "world", False), ("world",))], "Default Join"), STRING).output()
-
-    d2 \
-        .project(0, 1, 2) \
-        .map_partition(Verify([(1, 0.5, "hello"), (2, 0.4, "world")], "Project"), STRING).output()
-
-    d2 \
-        .union(d4) \
-        .map_partition(Verify2([(1, 0.5, "hello", True), (2, 0.4, "world", False), (1, 0.5, "hello", True), (1, 0.4, "hello", False), (1, 0.5, "hello", True), (2, 0.4, "world", False)], "Union"), STRING).output()
-
-    d4 \
-        .group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=False) \
-        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "AllGroupReduce"), STRING).output()
-
-    d4 \
-        .map(Id(), (INT, FLOAT, STRING, BOOL)).group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=True) \
-        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "ChainedGroupReduce"), STRING).output()
-
-    d4 \
-        .group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=True) \
-        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "CombineGroupReduce"), STRING).output()
-
-    d5 \
-        .group_by(2).sort_group(0, Order.DESCENDING).sort_group(1, Order.ASCENDING).reduce_group(GroupReduce3(), (FLOAT, FLOAT, INT), combinable=True) \
-        .map_partition(Verify([(4.3, 4.4, 1), (4.1, 4.1, 3)], "ChainedSortedGroupReduce"), STRING).output()
-
-    d4 \
-        .co_group(d5).where(0).equal_to(2).using(CoGroup(), ((INT, FLOAT, STRING, BOOL), (FLOAT, FLOAT, INT))) \
-        .map_partition(Verify([((1, 0.5, "hello", True), (4.4, 4.3, 1)), ((1, 0.4, "hello", False), (4.3, 4.4, 1))], "CoGroup"), STRING).output()
+    env.from_elements("hello", "world")\
+        .map(Id(), STRING).map_partition(Verify(["hello", "world"], "String"), STRING).output()
 
     #Custom Serialization
     class Ext(MapPartitionFunction):
@@ -285,6 +128,105 @@ if __name__ == "__main__":
         .map(Id(), CUSTOM).map_partition(Ext(), INT) \
         .map_partition(Verify([2, 4], "CustomTypeSerialization"), STRING).output()
 
+    #Map
+    class Mapper(MapFunction):
+        def map(self, value):
+            return value * value
+    d1 \
+        .map((lambda x: x * x), INT).map(Mapper(), INT) \
+        .map_partition(Verify([1, 1296, 20736], "Map"), STRING).output()
+
+    #FlatMap
+    class FlatMap(FlatMapFunction):
+        def flat_map(self, value, collector):
+            collector.collect(value)
+            collector.collect(value * 2)
+    d1 \
+        .flat_map(FlatMap(), INT).flat_map(FlatMap(), INT) \
+        .map_partition(Verify([1, 2, 2, 4, 6, 12, 12, 24, 12, 24, 24, 48], "FlatMap"), STRING).output()
+
+    #MapPartition
+    class MapPartition(MapPartitionFunction):
+        def map_partition(self, iterator, collector):
+            for value in iterator:
+                collector.collect(value * 2)
+    d1 \
+        .map_partition(MapPartition(), INT) \
+        .map_partition(Verify([2, 12, 24], "MapPartition"), STRING).output()
+
+    #Filter
+    class Filter(FilterFunction):
+        def __init__(self, limit):
+            super(Filter, self).__init__()
+            self.limit = limit
+
+        def filter(self, value):
+            return value > self.limit
+    d1 \
+        .filter(Filter(5)).filter(Filter(8)) \
+        .map_partition(Verify([12], "Filter"), STRING).output()
+
+    #Reduce
+    class Reduce(ReduceFunction):
+        def reduce(self, value1, value2):
+            return value1 + value2
+
+    class Reduce2(ReduceFunction):
+        def reduce(self, value1, value2):
+            return (value1[0] + value2[0], value1[1] + value2[1], value1[2], value1[3] or value2[3])
+    d1 \
+        .reduce(Reduce()) \
+        .map_partition(Verify([19], "AllReduce"), STRING).output()
+    d4 \
+        .group_by(2).reduce(Reduce2()) \
+        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "CombineReduce"), STRING).output()
+    d4 \
+        .map(Id(), (INT, FLOAT, STRING, BOOL)).group_by(2).reduce(Reduce2()) \
+        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "ChainedReduce"), STRING).output()
+
+    #GroupReduce
+    class GroupReduce(GroupReduceFunction):
+        def reduce(self, iterator, collector):
+            if iterator.has_next():
+                i, f, s, b = iterator.next()
+                for value in iterator:
+                    i += value[0]
+                    f += value[1]
+                    b |= value[3]
+                collector.collect((i, f, s, b))
+
+    class GroupReduce2(GroupReduceFunction):
+        def reduce(self, iterator, collector):
+            for value in iterator:
+                collector.collect(value)
+
+    class GroupReduce3(GroupReduceFunction):
+        def reduce(self, iterator, collector):
+            collector.collect(iterator.next())
+
+        def combine(self, iterator, collector):
+            if iterator.has_next():
+                v1 = iterator.next()
+            if iterator.has_next():
+                v2 = iterator.next()
+            if v1[0] < v2[0]:
+                collector.collect(v1)
+            else:
+                collector.collect(v2)
+    d4 \
+        .group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=False) \
+        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "AllGroupReduce"), STRING).output()
+    d4 \
+        .map(Id(), (INT, FLOAT, STRING, BOOL)).group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=True) \
+        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "ChainedGroupReduce"), STRING).output()
+    d4 \
+        .group_by(2).reduce_group(GroupReduce(), (INT, FLOAT, STRING, BOOL), combinable=True) \
+        .map_partition(Verify([(3, 1.4, "hello", True), (2, 0.4, "world", False)], "CombineGroupReduce"), STRING).output()
+    d5 \
+        .group_by(2).sort_group(0, Order.DESCENDING).sort_group(1, Order.ASCENDING).reduce_group(GroupReduce3(), (FLOAT, FLOAT, INT), combinable=True) \
+        .map_partition(Verify([(4.3, 4.4, 1), (4.1, 4.1, 3)], "ChainedSortedGroupReduce"), STRING).output()
+
+    #Execution
     env.set_degree_of_parallelism(1)
 
     env.execute(local=True)
